stub for notifications page
