+++ assets/js/components/notifications/Notifications.jsx
@@ -1,79 +1,24 @@
 import React, { Component } from 'react'
-import NotificationsMenu from './NotificationsMenu'
+import DashboardLayout from '../common/DashboardLayout'
 
 // MUI
-import IconButton from '@material-ui/core/IconButton'
-import SmallBadge from '../common/SmallBadge'
+import Typography from '@material-ui/core/Typography';
+import Paper from '@material-ui/core/Paper';
 
 // Icons
-import NotificationsIcon from '@material-ui/icons/Notifications'
-
-const initialNotifications = [
-  {
-    id: "123",
-    title: "Gateway Pending Confirmation",
-    body: "Confirm your gateways",
-    category: "gateways",
-    createdAt: 1528487101,
-    url: "/gateways",
-  },
-  {
-    id: "123",
-    title: "Gateway Pending Confirmation",
-    body: "Confirm your gateways",
-    category: "gateways",
-    createdAt: 1528487101,
-    url: "/gateways",
-  }
-]
+import NotificationsIcon from '@material-ui/icons/Notifications';
 
 class Notifications extends Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      notifications: initialNotifications,
-      anchorEl: null
-    }
-
-    this.handleClick = this.handleClick.bind(this)
-    this.handleClose = this.handleClose.bind(this)
-  }
-
-  handleClose() {
-    this.setState({
-      anchorEl: null
-    })
-  }
-
-  handleClick(e) {
-    this.setState({
-      anchorEl: e.currentTarget
-    })
-  }
-
   render() {
-    const { notifications, anchorEl } = this.state
-
     return (
-      <div>
-        <IconButton onClick={this.handleClick} color="inherit">
-          {notifications.length > 0 ?
-            <SmallBadge badgeContent={notifications.length} color="secondary">
-              <NotificationsIcon />
-            </SmallBadge>
-            :
-            <NotificationsIcon />
-          }
-        </IconButton>
-
-        <NotificationsMenu
-          notifications={notifications}
-          open={Boolean(anchorEl)}
-          anchorEl={anchorEl}
-          onClose={this.handleClose}
-        />
-      </div>
+      <DashboardLayout title="All Notifications">
+        <Paper style={{textAlign: 'center', padding: '5em'}}>
+          <NotificationsIcon style={{width: 80, height: 80, color: "#e0e0e0"}} />
+          <Typography variant="display1" style={{color: "#e0e0e0"}}>
+            Notifications goes here
+          </Typography>
+        </Paper>
+      </DashboardLayout>
     )
   }
 }
